requirements.txt
absl-py==1.4.0
amqp==5.1.1
asgiref==3.7.2
astunparse==1.6.3
beautifulsoup4==4.12.2
billiard==4.1.0
blinker==1.6.2
cachetools==5.3.1
celery==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.5
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==41.0.2
deepface==0.0.79
defusedxml==0.7.1
Django==4.2.3
django-allauth==0.54.0
django-crispy-forms==1.14.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
filelock==3.12.2
Flask==2.3.2
flatbuffers==23.5.26
gast==0.4.0
gdown==4.7.1
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.56.0
gunicorn==20.1.0
h5py==3.9.0
idna==3.4
itsdangerous==2.1.2
jax==0.4.13
Jinja2==3.1.2
keras==2.12.0
kombu==5.3.1
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.3
ml-dtypes==0.2.0
mtcnn==0.1.1
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
prompt-toolkit==3.0.39
protobuf==4.23.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.7.0
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.8
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
retina-face==0.0.13
rsa==4.9
scipy==1.11.1
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
tensorboard==2.12.3
tensorboard-data-server==0.7.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.3.0
tqdm==4.65.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
vine==5.0.0
wcwidth==0.2.6
Werkzeug==2.3.6
wrapt==1.14.1
